dev: add macOS support for devenv.nix
Add platform detection and conditional configuration to support native
development on macOS (aarch64-darwin) while preserving full Linux
functionality.

- Add platform detection using builtins.currentSystem for overlays
- Make packages conditional: patchelf, gcc, Linaro toolchain Linux-only
- Add macOS-specific pkg-config CFLAGS gathering for MuPDF build
- Make target directory platform-aware (Darwin vs Linux)
- Disable git hooks on macOS due to GDB dependency chain
- Fix statvfs type cast in ota/client.rs for cross-platform support
- Add contributor documentation for devenv setup with macOS notes

+++ crates/core/src/ota/client.rs
@@ -566,7 +566,7 @@ fn check_disk_space(path: &str) -> Result<(), OtaError> {
     use nix::sys::statvfs::statvfs;
 
     let stat = statvfs(path)?;
-    let available_mb = (stat.blocks_available() * stat.block_size()) / (1024 * 1024);
+    let available_mb = (stat.blocks_available() as u64 * stat.block_size() as u64) / (1024 * 1024);
     info!(
         "[OTA] Available disk space in {}: {} MB",
         path, available_mb
